Big cleanup of the module/include structure of the whole program Added a flac entry for the Rack mime_type method Most of the top-level logic in streamroller.rb moved to the initialization method of StreamRoller instead. Some methods moved out of Utils and into more specific files
abstracthandler, toolmanager, requestrouter, mp3passthrough, and flac_to_mp3 are all new files that define/utilize a file handler API for handling and serving audio files to the end user. This replaces a large sphagetti mess that used to be in the main music serving portion of the Sinatra StreamRoller object.